refactor: inline single-caller helper functions in tool-use flow
- Inline applyFollowUpMessage() - 3-line function called once
- Inline prepareInitialState() - ~55-line function called once from PrepareNode

The functions were moved directly into their single call sites, eliminating
abstraction overhead and making the code flow easier to follow. ~45 net lines
removed.
