+++ src/agent/implementations/flows/tooluse/index.ts
@@ -3,11 +3,6 @@
  *
  * Service interfaces and flow-first execution for tool-use agents.
  * The flow itself is in ToolUseRunFlow.ts (parent directory).
- *
- * Architecture (refactored - no closure wrappers):
- * - Helper functions are exported for direct use by nodes
- * - Services pass context values directly (resolvedTools, snapshot, etc.)
- * - Eliminates closure indirection for cleaner call stacks
  */
 
 export type {
@@ -19,9 +14,6 @@ export type {
 export {
   ToolUseFlowContext,
   type ToolUseFlowContextInit,
-  // Helper functions for direct use by nodes (no closure wrappers)
-  prepareInitialState,
-  applyFollowUpMessage,
 } from './ToolUseFlowContext';
 
 export {
